Made stake and unstake work
